Sweeping Changes
- Converted a bunch of Integers to Doubles for spacial calculations.
--- they need to be cast to (int), for visual effects; this is the price.
- Made the field into spacial and environmental data and rebuilt it.
- Added area outside of the zone of play, so the field isn't filling the whole window.
- Fleshed out some classes a bit more.,
- More stuff as well

+++ src/entities/Attacker.java
@@ -2,7 +2,7 @@ package entities;
 
 public class Attacker extends Player {
 
-    public Attacker(int number, int xPos, int yPos) {
+    public Attacker(int number, double xPos, double yPos) {
         super(number, xPos, yPos);
     }
